feat: adiciona regra de negócio no método delete do customerr controller

+++ src/controllers/customers-controller.js
@@ -55,7 +55,12 @@ class CustomersController {
 
     async delete(req, res) {
         const { id } = req.params
-
+        // Regra de negócio
+        const customer = await customersRepository.findById(id)
+        if (!customer) {
+            return res.status(404).json({ message: "Customer not found" })
+        }
+        
         const deletedCustomer = await customersRepository.delete(id)
 
         return res.json(deletedCustomer)
